rename commit. change stickings menu folder location
