--- src/components/Icons/SnareIcon.tsx
@@ -1,180 +1,0 @@
-const SnareIcon = () => {
-  return (
-    <svg
-      version="1.1"
-      id="Layer_1"
-      xmlns="http://www.w3.org/2000/svg"
-      xmlnsXlink="http://www.w3.org/1999/xlink"
-      x="0px"
-      y="0px"
-      width="100%"
-      viewBox="0 0 1024 1024"
-      enableBackground="new 0 0 1024 1024"
-      xmlSpace="preserve"
-    >
-      <path
-        fill="#000000"
-        opacity="1.000000"
-        stroke="none"
-        d="
-M326.273743,277.726379 
-	C368.465088,319.930603 410.412659,361.878632 452.311279,403.875549 
-	C455.534851,407.106689 458.733795,409.365204 463.635712,410.220581 
-	C474.524567,412.120636 483.424530,417.956970 489.672668,427.279694 
-	C491.370941,429.813629 493.138367,430.420410 496.249481,430.318146 
-	C507.400543,429.951660 518.575500,429.998230 529.730713,430.293884 
-	C532.548157,430.368530 534.311340,430.051300 535.983826,427.623077 
-	C543.058472,417.351624 552.813599,411.227203 565.252136,409.647491 
-	C568.193298,409.273987 569.943909,407.354034 571.827881,405.469757 
-	C622.612427,354.675293 673.432495,303.916199 724.155457,253.060303 
-	C731.759216,245.436569 740.116272,240.178589 751.379517,240.609589 
-	C763.856628,241.087036 773.144226,246.872330 778.817017,257.656433 
-	C784.491882,268.444580 783.903320,279.524323 777.406494,289.982941 
-	C775.753357,292.644196 773.624390,295.080261 771.399902,297.305695 
-	C725.449524,343.274658 679.452576,389.196991 633.474670,435.138397 
-	C632.438049,436.174133 631.117737,437.041687 630.843262,439.300659 
-	C637.529602,440.753845 644.296509,442.113556 651.008362,443.705597 
-	C666.587769,447.401001 681.933594,451.801544 696.185913,459.349365 
-	C702.265747,462.569183 707.925720,466.351776 712.509216,471.555176 
-	C716.956543,476.604095 719.393188,482.354431 719.388062,489.255219 
-	C719.325928,573.253540 719.350769,657.251831 719.342346,741.250183 
-	C719.341614,748.512634 717.326721,751.464478 710.462158,753.915161 
-	C686.067261,762.624329 660.790344,767.536743 635.255493,771.378296 
-	C584.027344,779.085266 532.456238,780.606750 480.812836,779.055115 
-	C432.022980,777.589172 383.536957,772.892029 336.150818,760.261108 
-	C329.728638,758.549316 323.382263,756.501953 317.104431,754.312805 
-	C308.843933,751.432129 306.825500,748.356628 306.805542,739.433777 
-	C306.752533,715.768188 306.702209,692.102539 306.694946,668.436951 
-	C306.677002,609.938110 306.873779,551.438171 306.529022,492.941345 
-	C306.455261,480.423615 311.596619,471.590607 321.296082,464.695312 
-	C334.004730,455.660645 348.557587,450.825806 363.361023,446.763458 
-	C373.757843,443.910309 384.332642,441.705780 395.154358,439.134064 
-	C394.629700,436.571960 392.535767,435.305206 390.980621,433.748169 
-	C345.165802,387.878754 299.313080,342.047089 253.487915,296.188019 
-	C244.525421,287.218872 241.300568,276.390472 244.501129,264.166870 
-	C247.738525,251.802551 256.111481,244.016708 268.530151,241.314484 
-	C280.252075,238.763870 290.596344,242.043686 299.148224,250.615173 
-	C308.094086,259.581543 317.060303,268.527588 326.273743,277.726379 
-M511.511932,452.312653 
-	C508.846588,452.312286 506.163208,452.508942 503.520050,452.266632 
-	C499.270203,451.877075 497.656616,453.324005 496.854187,457.852844 
-	C492.858521,480.404114 476.245178,494.401062 454.308533,494.353516 
-	C435.506775,494.312775 419.545441,482.417999 413.813263,463.840485 
-	C412.712555,460.273254 411.478882,459.271851 407.532806,459.933411 
-	C391.105499,462.687286 374.860657,466.100311 359.012115,471.270111 
-	C350.760010,473.961975 342.687653,477.110504 335.429108,481.951538 
-	C327.567322,487.194885 327.606445,490.156372 335.351074,495.398071 
-	C342.170410,500.013489 349.689789,503.172089 357.464966,505.710327 
-	C398.189148,519.004822 440.412140,522.844421 482.822113,524.496155 
-	C519.922363,525.941040 557.036865,525.086182 594.003845,520.770447 
-	C617.483826,518.029175 640.765259,514.311646 663.424377,507.454712 
-	C674.264648,504.174286 685.009766,500.421326 693.963623,493.082184 
-	C698.328430,489.504486 698.218872,487.564941 693.769470,484.040253 
-	C689.030579,480.286316 683.669739,477.619598 678.100830,475.337006 
-	C659.358093,467.654846 639.681946,463.638336 619.837097,460.123718 
-	C615.169922,459.297119 613.246948,459.885773 611.781372,464.830994 
-	C606.517578,482.592102 590.157227,494.308380 571.619873,494.354919 
-	C552.455444,494.403015 536.646423,482.927490 530.545044,465.205292 
-	C529.093933,460.990326 531.136780,454.733612 526.360901,452.817413 
-	C522.324158,451.197845 517.165588,452.374207 511.511932,452.312653 
-M604.291748,700.222290 
-	C609.630188,708.536499 611.995178,708.506226 617.509583,699.943115 
-	C643.120056,660.173279 668.597656,620.316895 694.433960,580.694397 
-	C698.431641,574.563477 697.412170,568.543579 697.513367,561.603943 
-	C645.938782,583.921631 591.702332,589.898438 536.386169,594.577942 
-	C559.263489,630.160583 581.589172,664.885254 604.291748,700.222290 
-M386.449219,580.273193 
-	C367.058319,575.885803 348.045044,570.282104 329.453064,561.692993 
-	C329.324280,569.204590 328.471558,575.508240 332.563324,581.767883 
-	C353.707672,614.114990 374.471283,646.710938 395.365906,679.221191 
-	C400.409729,687.068909 405.377258,694.966858 410.513458,702.753662 
-	C413.493469,707.271484 417.266235,707.229187 420.201630,702.832825 
-	C421.865967,700.340149 423.463928,697.802979 425.085632,695.281921 
-	C445.541718,663.483032 466.010559,631.692383 486.409607,599.856995 
-	C487.394928,598.319275 489.206635,596.919250 488.452148,594.457886 
-	C477.323212,593.417908 466.230804,592.441162 455.151459,591.333618 
-	C432.309387,589.050232 409.668579,585.494568 386.449219,580.273193 
-M338.522552,708.114136 
-	C335.580322,707.289734 333.142731,704.881470 329.303711,705.097046 
-	C329.303711,714.453186 329.297668,723.582092 329.308289,732.711060 
-	C329.310699,734.769226 330.704620,735.536743 332.431122,736.033630 
-	C336.584473,737.229309 340.692139,738.597717 344.876160,739.670410 
-	C373.931671,747.119568 403.524567,751.091003 433.353912,753.885254 
-	C460.072174,756.388062 486.852692,757.309021 513.624817,757.285217 
-	C550.060669,757.252869 586.455200,755.553101 622.645813,750.741150 
-	C646.569092,747.560364 670.228760,743.221924 693.421814,736.455444 
-	C696.745911,735.485718 697.819031,733.721985 697.717529,730.339905 
-	C697.492737,722.850769 697.676636,715.350037 697.620178,707.854675 
-	C697.611023,706.637207 698.204956,705.139282 696.410645,704.166504 
-	C687.256836,708.630676 677.665283,712.362366 667.791992,715.363159 
-	C619.964050,729.899597 570.782593,735.049744 521.077087,735.867493 
-	C482.762299,736.497925 444.606934,733.768982 406.817291,727.211487 
-	C383.690399,723.198364 360.880768,717.815125 338.522552,708.114136 
-M697.648743,524.680969 
-	C697.392517,522.601562 698.228088,520.397400 697.083496,518.471375 
-	C695.649963,518.008179 694.701050,518.851257 693.696533,519.353882 
-	C680.807434,525.803528 667.115295,529.955933 653.168762,533.276001 
-	C615.786499,542.174866 577.707458,545.645386 539.432129,546.945129 
-	C496.772095,548.393616 454.222717,546.655884 411.929840,540.433533 
-	C384.385864,536.381104 357.122345,531.262268 331.686462,519.219604 
-	C331.153168,518.967163 330.448181,519.077332 329.819214,519.017883 
-	C328.433563,536.431580 327.907593,534.727356 340.565155,541.933899 
-	C348.258057,546.313782 356.686127,549.106262 365.145630,551.649475 
-	C396.956543,561.213379 429.562683,566.454041 462.551117,569.819946 
-	C490.113068,572.632141 517.693115,573.621643 545.316162,571.547485 
-	C577.017761,569.167053 608.478882,564.906372 639.444763,557.559326 
-	C656.465393,553.520935 673.305359,548.861389 688.867920,540.484192 
-	C694.932068,537.219910 699.235474,533.205383 697.648743,524.680969 
-M560.785583,673.644165 
-	C545.095215,649.219666 529.404785,624.795166 513.135071,599.468872 
-	C489.201355,636.756531 465.968781,672.951843 442.169647,710.029785 
-	C489.841736,716.553040 536.354797,716.646362 584.021790,709.938232 
-	C576.017334,697.425720 568.602661,685.835083 560.785583,673.644165 
-M440.769897,466.733521 
-	C441.830414,467.538452 442.842438,468.418976 443.959351,469.136017 
-	C451.028992,473.674622 460.592133,473.145477 467.343201,467.882050 
-	C474.151581,462.574005 476.850586,453.369904 474.042725,445.035614 
-	C471.347229,437.034790 464.029358,431.400543 455.613556,431.406097 
-	C450.173462,431.409698 446.168427,429.157166 442.451569,425.434021 
-	C390.560577,373.454926 338.611023,321.534302 286.668243,269.606934 
-	C285.020935,267.960083 283.446198,266.182739 281.575958,264.827301 
-	C275.355896,260.319519 267.202118,263.492035 265.639740,270.999054 
-	C264.705994,275.485565 267.320404,278.596039 270.198334,281.479431 
-	C285.846527,297.157227 301.501099,312.828674 317.177460,328.478241 
-	C354.305206,365.542389 391.434448,402.605042 428.603729,439.627502 
-	C431.781891,442.793152 434.393402,446.158142 434.237854,450.856201 
-	C434.038147,456.888031 436.320404,461.865967 440.769897,466.733521 
-M682.181824,326.681915 
-	C649.322205,359.543396 616.438171,392.380554 583.635925,425.299103 
-	C579.795166,429.153412 575.663086,431.364105 570.028076,431.432617 
-	C559.197937,431.564240 550.384338,441.235870 550.564270,452.175995 
-	C550.740295,462.884644 560.118530,472.227631 570.837097,472.372650 
-	C581.604187,472.518372 591.458984,463.527008 591.548279,452.816498 
-	C591.593872,447.348358 593.573547,443.374939 597.330566,439.629761 
-	C637.201538,399.884216 676.984558,360.050507 716.793335,320.242584 
-	C730.102112,306.934052 743.413025,293.627655 756.728027,280.325287 
-	C759.931030,277.125336 761.310181,273.496307 759.621887,269.026337 
-	C758.155396,265.143555 755.102234,263.320770 751.269348,262.797821 
-	C747.032776,262.219788 744.102173,264.727051 741.322266,267.513153 
-	C721.792480,287.086334 702.233521,306.630463 682.181824,326.681915 
-M685.978760,684.493591 
-	C687.005249,683.940674 688.012634,683.348145 689.061462,682.841370 
-	C695.154480,679.897461 698.113586,675.667114 697.832581,668.296997 
-	C697.236511,652.660400 697.660706,636.985352 697.624329,621.326172 
-	C697.621704,620.175598 698.065247,618.845886 696.860535,617.574768 
-	C679.685913,644.357239 662.669434,670.893066 644.953247,698.520081 
-	C659.763428,695.027039 672.891846,690.997620 685.978760,684.493591 
-M348.352753,647.095398 
-	C342.349396,637.803894 336.346008,628.512451 330.204987,619.007874 
-	C328.800964,620.728760 329.221863,622.106689 329.220062,623.367920 
-	C329.197571,639.196045 329.424927,655.028076 329.118317,670.850098 
-	C329.020081,675.917908 331.056641,678.869995 335.059662,681.368164 
-	C345.059753,687.609070 356.044098,691.423096 367.218750,694.788513 
-	C371.410645,696.051025 375.531342,697.764221 380.944885,697.933105 
-	C369.913177,680.738281 359.320374,664.227539 348.352753,647.095398 
-z"
-      />
-    </svg>
-  );
-};
-export default SnareIcon;
