+++ src/components/StickingsMenu/MenuForms/MenuInputs/RowInput.tsx
@@ -1,4 +1,4 @@
-import { isRowChecked } from '../../../../../lib/uiHelpers/menuHelpers';
+import { isRowChecked } from '../../../../lib/uiHelpers/menuHelpers';
 
 interface RowInputProps {
   inputType: string;
