+++ src/components/StickingsMenu/StickingsMenu.tsx
