+++ src/pages/TripletNotesMenu/TripletNotesMenu.tsx
@@ -1,5 +1,5 @@
 import Staff from '../../components/Staff/Staff';
-import StickingsMenu from '../../components/Menu/StickingsMenu/StickingsMenu';
+import StickingsMenu from '../../components/StickingsMenu/StickingsMenu';
 
 import { tripletPermutations } from '../../lib/utils/permutations';
 import getTripletNotesArray from '../../lib/utils/staffUtils/getTripletNotesArray';
